Enhance donation features and optimize frontend/backend code (#11)
* Backend:Code formatted and laravel pint installed

* Frontend:Code formatted, eslint and prettier added

* Frontend:Pinia issue solved

* Frontend:Added donation button

* Backend:Tailwindcss issue fixed

* Frontend:Home page and donation button issue fixed

* Frontend: Create Donation Formula form develop

* Frontend:: Update donation formula modal

* Frontend:: Remove Donation Formula Page option from new article page

* Frontend:: Main header remove and show default article

* Donation formula

* Donation formula new login added

* Donation formula UI updated

* Formula share button created

* Code optimize and formatted

* Fixed

* Merge branch 'main' into develop

+++ backend/database/migrations/2025_04_23_064923_create_payments_table.php
@@ -11,17 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('payments', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->nullable()->constrained()->onDelete('set null');
-            $table->string('payment_id');
-            $table->string('customer_id');
-            $table->decimal('amount', 10, 2);
-            $table->string('currency')->default('usd');
-            $table->string('status');
-            $table->timestamps();
-
-        });
+        if (! Schema::hasTable('payments')) {
+            Schema::create('payments', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->nullable()->constrained()->onDelete('set null');
+                $table->string('payment_id');
+                $table->string('customer_id');
+                $table->decimal('amount', 10, 2);
+                $table->string('currency')->default('usd');
+                $table->string('status');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
