+++ backend/database/migrations/2025_05_01_070135_create_payment_logs_table.php
@@ -8,14 +8,16 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('payment_logs', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('cause_id')->constrained()->onDelete('cascade');
-            $table->foreignId('ngo_id')->constrained()->onDelete('cascade');
-            $table->foreignId('payment_id')->constrained()->onDelete('cascade');
-            $table->decimal('amount', 10, 2);
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('payment_logs')) {
+            Schema::create('payment_logs', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('cause_id')->constrained()->onDelete('cascade');
+                $table->foreignId('ngo_id')->constrained()->onDelete('cascade');
+                $table->foreignId('payment_id')->constrained()->onDelete('cascade');
+                $table->decimal('amount', 10, 2);
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
